Fixed some implicit cast warnings.
git-svn-id: https://svn.outpostuniverse.org:8443/svn/outposthd@244 bdf9137d-f3be-7946-88fe-7b0bbfdbf31a

+++ src/Constants/Numbers.h
@@ -3,7 +3,7 @@
 
 namespace constants
 {
-	const float FADE_SPEED = 175.0f;
+	const int FADE_SPEED = 175;
 
 	const int PLANET_ANIMATION_SPEED = 65;
